- update documentation, closes #73
