+++ examples/README.md
@@ -1,0 +1,29 @@
+# GoDS (Go Data Structures)
+
+Various examples on how to use data structures.
+
+## Examples
+
+- [ArrayList](https://github.com/emirpasic/gods/blob/master/examples/arraylist/arraylist.go)
+- [ArrayStack](https://github.com/emirpasic/gods/blob/master/examples/arraystack/arraystack.go)
+- [AVLTree](https://github.com/emirpasic/gods/blob/master/examples/avltree/avltree.go)
+- [BinaryHeap](https://github.com/emirpasic/gods/blob/master/examples/binaryheap/binaryheap.go)
+- [BTree](https://github.com/emirpasic/gods/blob/master/examples/btree/btree.go)
+- [Custom Comparator](https://github.com/emirpasic/gods/blob/master/examples/customcomparator/customcomparator.go)
+- [DoublyLinkedList](https://github.com/emirpasic/gods/blob/master/examples/doublylinkedlist/doublylinkedlist.go)
+- [EnumerableWithIndex](https://github.com/emirpasic/gods/blob/master/examples/enumerablewithindex/enumerablewithindex.go)
+- [EnumerableWithKey](https://github.com/emirpasic/gods/blob/master/examples/enumerablewithkey/enumerablewithkey.go)
+- [HashBidiMap](https://github.com/emirpasic/gods/blob/master/examples/hashbidimap/hashbidimap.go)
+- [HashMap](https://github.com/emirpasic/gods/blob/master/examples/hashmap/hashmap.go)
+- [HashSet](https://github.com/emirpasic/gods/blob/master/examples/hashset/hashset.go)
+- [IteratorWithIndex](https://github.com/emirpasic/gods/blob/master/examples/iteratorwithindex/iteratorwithindex.go)
+- [iteratorwithkey](https://github.com/emirpasic/gods/blob/master/examples/iteratorwithkey/iteratorwithkey.go)
+- [IteratorWithKey](https://github.com/emirpasic/gods/blob/master/examples/linkedliststack/linkedliststack.go)
+- [RedBlackTree](https://github.com/emirpasic/gods/blob/master/examples/redblacktree/redblacktree.go)
+- [RedBlackTreeExtended](https://github.com/emirpasic/gods/blob/master/examples/redblacktreeextended/redblacktreeextended.go)
+- [Serialization](https://github.com/emirpasic/gods/blob/master/examples/serialization/serialization.go)
+- [SinglyLinkedList](https://github.com/emirpasic/gods/blob/master/examples/singlylinkedlist/singlylinkedlist.go)
+- [Sort](https://github.com/emirpasic/gods/blob/master/examples/sort/sort.go)
+- [TreeBidiMap](https://github.com/emirpasic/gods/blob/master/examples/treebidimap/treebidimap.go)
+- [TreeMap](https://github.com/emirpasic/gods/blob/master/examples/treemap/treemap.go)
+- [TreeSet](https://github.com/emirpasic/gods/blob/master/examples/treeset/treeset.go)
