--- examples/arraylist.go
@@ -1,31 +1,0 @@
-// Copyright (c) 2015, Emir Pasic. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package examples
-
-import (
-	"github.com/emirpasic/gods/lists/arraylist"
-	"github.com/emirpasic/gods/utils"
-)
-
-// ArrayListExample to demonstrate basic usage of ArrayList
-func ArrayListExample() {
-	list := arraylist.New()
-	list.Add("a")                         // ["a"]
-	list.Add("c", "b")                    // ["a","c","b"]
-	list.Sort(utils.StringComparator)     // ["a","b","c"]
-	_, _ = list.Get(0)                    // "a",true
-	_, _ = list.Get(100)                  // nil,false
-	_ = list.Contains("a", "b", "c")      // true
-	_ = list.Contains("a", "b", "c", "d") // false
-	list.Swap(0, 1)                       // ["b","a",c"]
-	list.Remove(2)                        // ["b","a"]
-	list.Remove(1)                        // ["b"]
-	list.Remove(0)                        // []
-	list.Remove(0)                        // [] (ignored)
-	_ = list.Empty()                      // true
-	_ = list.Size()                       // 0
-	list.Add("a")                         // ["a"]
-	list.Clear()                          // []
-}
