+++ examples/btree/btree.go
@@ -1,0 +1,59 @@
+// Copyright (c) 2015, Emir Pasic. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package main
+
+import (
+	"fmt"
+	"github.com/emirpasic/gods/trees/btree"
+)
+
+// BTreeExample to demonstrate basic usage of BTree
+func main() {
+	tree := btree.NewWithIntComparator(3) // empty (keys are of type int)
+
+	tree.Put(1, "x") // 1->x
+	tree.Put(2, "b") // 1->x, 2->b (in order)
+	tree.Put(1, "a") // 1->a, 2->b (in order, replacement)
+	tree.Put(3, "c") // 1->a, 2->b, 3->c (in order)
+	tree.Put(4, "d") // 1->a, 2->b, 3->c, 4->d (in order)
+	tree.Put(5, "e") // 1->a, 2->b, 3->c, 4->d, 5->e (in order)
+	tree.Put(6, "f") // 1->a, 2->b, 3->c, 4->d, 5->e, 6->f (in order)
+	tree.Put(7, "g") // 1->a, 2->b, 3->c, 4->d, 5->e, 6->f, 7->g (in order)
+
+	fmt.Println(tree)
+	// BTree
+	//         1
+	//     2
+	//         3
+	// 4
+	//         5
+	//     6
+	//         7
+
+	_ = tree.Values() // []interface {}{"a", "b", "c", "d", "e", "f", "g"} (in order)
+	_ = tree.Keys()   // []interface {}{1, 2, 3, 4, 5, 6, 7} (in order)
+
+	tree.Remove(2) // 1->a, 3->c, 4->d, 5->e, 6->f (in order)
+	fmt.Println(tree)
+	// BTree
+	//     1
+	//     3
+	// 4
+	//     5
+	//     6
+
+	tree.Clear() // empty
+	tree.Empty() // true
+	tree.Size()  // 0
+
+	// Other:
+	tree.Height()     // gets the height of the tree
+	tree.Left()       // gets the left-most (min) node
+	tree.LeftKey()    // get the left-most (min) node's key
+	tree.LeftValue()  // get the left-most (min) node's value
+	tree.Right()      // get the right-most (max) node
+	tree.RightKey()   // get the right-most (max) node's key
+	tree.RightValue() // get the right-most (max) node's value
+}
