+++ examples/treeset/treeset.go
@@ -1,0 +1,23 @@
+// Copyright (c) 2015, Emir Pasic. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package main
+
+import "github.com/emirpasic/gods/sets/treeset"
+
+// TreeSetExample to demonstrate basic usage of TreeSet
+func main() {
+	set := treeset.NewWithIntComparator() // empty
+	set.Add(1)                            // 1
+	set.Add(2, 2, 3, 4, 5)                // 1, 2, 3, 4, 5 (in order, duplicates ignored)
+	set.Remove(4)                         // 1, 2, 3, 5 (in order)
+	set.Remove(2, 3)                      // 1, 5 (in order)
+	set.Contains(1)                       // true
+	set.Contains(1, 5)                    // true
+	set.Contains(1, 6)                    // false
+	_ = set.Values()                      // []int{1,5} (in order)
+	set.Clear()                           // empty
+	set.Empty()                           // true
+	set.Size()                            // 0
+}
